% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biblesearch_versions.R
\name{get_biblesearch_versions}
\alias{get_biblesearch_versions}
\title{Get Available BibleSearch Versions}
\usage{
get_biblesearch_versions()
}
\value{
data frame of versions
}
\description{
Get Available BibleSearch Versions
}
\examples{
get_biblesearch_versions()
}
