% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_biblesearch_key.R
\name{set_biblesearch_key}
\alias{set_biblesearch_key}
\title{Set API Key}
\usage{
set_biblesearch_key(key)
}
\arguments{
\item{key}{}
}
\description{
Set API Key
}
\examples{
set_biblesearch_key("fdsdffdssd4324213234")
}
